+++ test/core/non_ansi_port_decl.v
@@ -1,0 +1,2 @@
+`define REF
+`include "non_ansi_port_decl.sv"