+++ test/core/non_ansi_port_decl_order.v
@@ -1,0 +1,9 @@
+module mod(a, b, c, d);
+    input wire a;
+    output reg b;
+    always @*
+        b = ~a;
+    input wire [7:0] c;
+    output wire [31:0] d;
+    assign d = 10 + c;
+endmodule