+++ test/error/port_not_in_header.sv
@@ -1,0 +1,4 @@
+// pattern: declared ports w, z are not in the port list of top
+module top(x, y);
+    output w, x, y, z;
+endmodule