+++ test/error/port_unpacked_second.sv
@@ -1,0 +1,5 @@
+// pattern: declarations `output x` and `wire x \[1:0\]` are incompatible due to different unpacked dimensions
+module top(x);
+    output x;
+    wire x [1:0];
+endmodule